requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.22.0
attrs==17.2.0
autobahn==17.9.2
Automat==0.6.0
billiard==3.3.0.23
celery==3.1.25
certifi==2017.7.27.1
cffi==1.10.0
channels==1.1.6
chardet==3.0.4
constantly==15.1.0
cryptography==2.0.3
daphne==1.3.0
Django==1.11.5
django-allauth==0.33.0
django-celery==3.2.1
django-channels-presence==0.0.8
djangorestframework==3.6.4
enum34==1.1.6
geoip2==2.5.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.18
jsonfield==2.0.2
kombu==3.0.37
maxminddb==1.3.0
msgpack-python==0.4.8
oauthlib==2.0.4
olefile==0.44
Pillow==4.3.0
psutil==5.4.0
psycopg2==2.7.3.1
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
pyOpenSSL==17.3.0
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
service-identity==17.0.0
six==1.10.0
Twisted==17.5.0
txaio==2.8.2
urllib3==1.22
vine==1.1.4
zope.interface==4.4.2
